+++ sql/dolibarr_18.0.0-19.0.0.sql
@@ -1,0 +1,5 @@
+--
+-- Script run when an upgrade of Dolibarr is done. Upgrade Dolibarr version 18.0.0 to 19.0.0.
+--
+
+update llx_actioncomm set elementtype = 'funding@funding' where code LIKE'%AC_FUNDING%' and elementtype = 'funding';